# src/openapi/admin-services.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Admin Services API",
    "version": "1.1.0",
    "description": "Admin endpoints for managing Services, analytics, and exports. Includes error schemas and HEAD endpoint."
  },
  "servers": [
    { "url": "/api" }
  ],
  "paths": {
...
    "/admin/services/stats": {
      "get": {
        "summary": "Get services analytics/stats",
        "parameters": [ { "name": "range", "in": "query", "schema": { "type": "string", "enum": ["7d","30d","90d"] } } ],
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "Unauthorized", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "403": { "description": "Forbidden", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "message": { "type": "string" },
          "details": {}
        },
        "required": ["code","message"]
      },
...
    }
  }
}